Add casia-webface to .gitignore and refactor classification

+++ UMUT/preprocess/classification/folder_operations_c.py
@@ -1,6 +1,6 @@
 import os
 import shutil
-
+import headpose_operations_c
 
 def delete_old_group_folders(folder_path: str) -> None:
     """
@@ -12,6 +12,8 @@ def delete_old_group_folders(folder_path: str) -> None:
         Returns:
             None
     """
+    if not os.path.exists(folder_path):
+        return
     for folder_name in os.listdir(folder_path):
         if "group" in folder_name:
             shutil.rmtree(os.path.join(folder_path, folder_name))
@@ -20,6 +22,8 @@ def delete_old_group_folders(folder_path: str) -> None:
 def get_img_file_count(files: list) -> int:
     count = 0
     for file in files:
+        if "group" in file or "frontal" in file:
+            continue
         count = count+1 if file.endswith(".jpg") else count 
     return count
 
@@ -35,18 +39,29 @@ def build_group_folders(groups: list, print_flag: bool = False) -> None:
             None
     """
     for group_index,group in enumerate(groups):
+        
+        print("Group: ",group_index)
+        headpose_response = headpose_operations_c.get_most_frontal_face(group)
+
         for obj in group:
             # Create the folder path
             group_slices = obj.path.split(os.sep)
             group_slices[-1] = f"group_{group_index}/{group_slices[-1]}"
+            #group_slices.pop(-2)
             group_folder_path = os.path.normpath('/'.join(group_slices))
+            group_folder_path = group_folder_path.replace(obj.source_path, obj.target_path)
+            if False:#not headpose_response["valid_data"]:
+                continue
+
+            frontal_img = headpose_response["frontal_img"]
             
-            if not os.path.exists(group_folder_path):
-                os.makedirs(os.path.dirname(group_folder_path), exist_ok=True)
-            else:
+            if os.path.exists(group_folder_path):
                 shutil.rmtree( os.path.dirname(group_folder_path) )
-                os.makedirs(os.path.dirname(group_folder_path), exist_ok=True)
-            shutil.copy(obj.path, group_folder_path)
             
-            print(obj.path) if print_flag else None
+            os.makedirs(os.path.dirname(group_folder_path), exist_ok=True)
+            shutil.copy(obj.path, group_folder_path)
+            os.makedirs(os.path.join(os.path.dirname(group_folder_path), "frontal"), exist_ok=True)
+            shutil.copy(frontal_img.path, os.path.join(os.path.dirname(group_folder_path), "frontal", os.path.basename(frontal_img.path)) )
+            
+            print(obj.path,"to",group_folder_path) if print_flag else None
         print("\n") if print_flag else None
